Поправил сервис по работе с клиентами, вынес интерфейс сервиса и сделал его реализацию
