+++ src/main/java/ru/bakulin/daily_booking_service/mapper/ClientMapper.java
@@ -2,10 +2,11 @@ package ru.bakulin.daily_booking_service.mapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
+import org.mapstruct.MappingConstants;
 import ru.bakulin.daily_booking_service.dto.ClientDto;
 import ru.bakulin.daily_booking_service.entity.Client;
 
-@Mapper
+@Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
 public interface ClientMapper {
 
   @Mapping(target = "id", ignore = true)
